now can get actor bio from database
Need to make sure to do a delete before every insert still for episodes, cast_and_crew and also actor bios.
Do this to make sure that we dont have duplicates; not sure if MySQL can handle that on its own
Consider adding an extra unique field which is a hash of relevant data, this would prevent duplicates.