new_thetvdb_interface is ready!
new_thetvdb_interface is complete and can now be used, though it should
considered of 'beta' integrity.
Abstracted the storing of episodes and series into seperate functions,
both of which are called from a parent to handle a series with episodes
when processed with populate_results(), so that the store procedures
could be used in the udpate_db() method as well.
Minor updates.