wrote same_file?() and collect_duplicates()
same_file?() checks if two files are identical by
first checking that the file sizes match, and then
by progressively hashing the file until the hashes
differ or until EOF is reached.  The difference
between this and File.compare is the return value,
this returns an array with TRUE or FALSE as it's
first element and nil or the hash value of the file
as the second element.
collect_duplicates() iterates through a list of files,
looking for duplicates.  It returns a hash which has
hash values associated with an array of the files with
that hash value.