MovieInfo.merge and MovieInfo.clearMerges()
MovieInfo.merge provides the ability to store up to 5 items, in
a round robin fashion.  On the sixth merge, you will be overwriting
the first available slot, and so on.  It is meant to provide the
ability to pass along temporary information, such as how a file
was matched.

MovieInfo.clearMerges is self explanitory, it clears all slots and
resets the index to zero.