created Internals module, and moved matching to fuzzyMatch()
MediaManager::Internals module is meant to provide access to some
of the internal functions of MediaManager for development and
debugging.  So far only one function is in there, fuzzyMatch().
Will add them as they need to be debugged/troubleshooted.

Segregated name and epName matching into a function that matches
strings using various fuzzy methods.  This way it can be used to
match Titles,  and Episode Names, and anything else one could possibly
ever want to match.